+++ WeaselHost.Web/wwwroot/assets/LogPanel-CKiBB2IR.js
@@ -1,4 +1,4 @@
-import{c as N,r as n,u as m,m as P,b as v,j as t,x as R,n as S}from"./index-DKeW84DF.js";/**
+import{c as N,r as n,u as m,m as P,b as v,j as t,x as R,o as S}from"./index-eo1TGWw1.js";/**
  * @license lucide-react v0.439.0 - ISC
  *
  * This source code is licensed under the ISC license.